+++ modules/compute/aws-batch/app.py
@@ -12,6 +12,7 @@ from stack import AwsBatch
 project_name = os.getenv("SEEDFARMER_PROJECT_NAME", "")
 deployment_name = os.getenv("SEEDFARMER_DEPLOYMENT_NAME", "")
 module_name = os.getenv("SEEDFARMER_MODULE_NAME", "")
+partition = os.getenv("AWS_PARTITION", "aws")
 
 if len(f"{project_name}-{deployment_name}") > 36:
     raise ValueError("This module cannot support a project+deployment name character length greater than 35")
@@ -54,6 +55,7 @@ app = App()
 stack = AwsBatch(
     scope=app,
     id=f"{project_name}-{deployment_name}-{module_name}",
+    partition=partition,
     project_name=project_name,
     deployment_name=deployment_name,
     module_name=module_name,