+++ modules/compute/eks/tests/test_stack.py
@@ -14,7 +14,7 @@ from moto import mock_ec2, mock_eks
 def stack_defaults():
     os.environ["CDK_DEFAULT_ACCOUNT"] = "111111111111"
     os.environ["CDK_DEFAULT_REGION"] = "us-east-1"
-
+    os.environ["AWS_PARTITION"] = "aws"
     # Unload the app import so that subsequent tests don't reuse
 
     if "stack" in sys.modules:
@@ -80,6 +80,7 @@ def test_synthesize_stack(stack_defaults):
     eks_stack = stack.Eks(
         scope=app,
         id=f"{project_name}-{dep_name}-{mod_name}",
+        partition="aws",
         project_name=project_name,
         deployment_name=dep_name,
         module_name=mod_name,