+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### **Added**
 
+- added partition support for local zone poc modules
+  - aws-batch
+  - eks
+  - emr-on-eks
+  - buckets
+
 ### **Changed**
 
 ### **Removed**